Server/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-render-partial==0.4
djangorestframework==3.12.2
importlib-metadata==3.7.0
Markdown==3.3.4
numpy==1.20.1
psycopg2-binary==2.8.6
pytz==2021.1
Shapely==1.7.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
